Extract local type from function that creates it
