+++ src/cli/facade/clean.ts
@@ -2,7 +2,7 @@ import readCb from 'read';
 import { fromPromise } from 'universalify';
 import { promisify } from 'util';
 import strings from '../../resources/index';
-import { Local } from '../../types';
+import type { Local } from '../domain/local';
 import { Logger } from '../logger';
 
 const read = promisify(readCb);
