+++ src/cli/facade/mock.ts
@@ -1,6 +1,6 @@
 import { fromPromise } from 'universalify';
 import strings from '../../resources/index';
-import { Local } from '../../types';
+import type { Local } from '../domain/local';
 import { Logger } from '../logger';
 
 const mock = ({ local, logger }: { local: Local; logger: Logger }) =>
